physiotwin-ai/backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.32.1
sqlalchemy==2.0.36
pydantic==2.10.4
pydantic-settings==2.6.1
email-validator==2.2.0
passlib==1.7.4
python-jose==3.3.0
python-multipart==0.0.20
alembic==1.14.0
reportlab==4.2.5


